ggplot2 theme-like object for forcing axes to sit at (0, 0)

This is a ggplot2 theme designed for graphs in which axes are fixed at (0,0) or some other numbers. This is technically not a theme as it includes geometries as well. Use theme() to modify afterwards.

axis_zero(
  x_axis = 0,
  y_axis = 0,
  quadrants = 1,
  x_min = -10,
  y_min = -10,
  x_max = 10,
  y_max = 10,
  x_title = "X",
  y_title = "Y",
  title_distance = 2,
  title_distance_x = title_distance,
  title_distance_y = title_distance,
  x_title_loc = c(x_max, x_axis - title_distance_x),
  y_title_loc = c(y_axis - title_distance_y, y_max),
  title_opts = list(hjust = 1),
  title_opts_x = title_opts,
  title_opts_y = append(title_opts, list(angle = 90)),
  digits = 0,
  x_digits = digits,
  y_digits = digits,
  axis_distance = 0.2,
  axis_distance_x = axis_distance,
  axis_distance_y = axis_distance,
  axis_opts = list(),
  axis_opts_x = axis_opts,
  axis_opts_y = axis_opts,
  axis_format_x = function(x) scales::number(x, accuracy = 10^(-x_digits)),
  axis_format_y = function(x) scales::number(x, accuracy = 10^(-y_digits)),
  x_breaks = c(x_max * 0.9 * 0.25 * 1:4, x_min * 0.9 * 0.25 * 1:4) %>% round(digits =
    x_digits),
  y_breaks = c(y_max * 0.9 * 0.25 * 1:4, y_min * 0.9 * 0.25 * 1:4) %>% round(digits =
    y_digits),
  label_origin = TRUE,
  axis_text_size = 12,
  line_opts = list(),
  line_opts_x = NULL,
  line_opts_y = NULL,
  base_theme = ggplot2::theme_void,
  base_size = 18,
  base_family = "",
  base_line_size = base_size/10,
  base_rect_size = base_size/10
)

Arguments

x_axis
The value of y at which the x-axis will be drawn.

y_axis
The value of x at which the y-axis will be drawn.

x_min, x_max, y_min, y_max
The edges of the graph on the x and y axes, for the purposes of figuring out how far to draw the axes and place the axis titles. Only relevant if positive/negative quadrants included, respectively.

title_distance, title_distance_x, title_distance_y
The distance between the axis line and the title for that axis.

x_title_loc, y_title_loc
The locations of the axis titles set explicitly, in terms of actual x and y values, overriding title_distance.

title_opts, title_opts_x, title_opts_y
A list of options to be passed to ggplot2::annotate for the axis titles, for both axes or only x or y specifically.

digits, x_digits, y_digits
Number of digits after the decimal place to include in axis labels.

axis_distance, axis_distance_x, axis_distance_y
The distance between the axis line and the axis labels.

axis_opts, axis_opts_x, axis_opts_y
A list of options to be passed to ggplot2::annotate for the axis labels, for both axes or only x or y specifically.

axis_format_x, axis_format_y
The labeling function from the **scales** package to use to format the axis labels.

x_breaks, y_breaks
The axis labels to include. Including the origin values in either of these is ill-advised, use label_origin instead.

label_origin
Include an axis marker for the origin. Set to a character variable to specify the value.

axis_text_size
The font size of the axis labels.

line_opts, line_opts_x, line_opts_y
A list of options to be passed to ggplot2::geom_segment for the axis lines, for both axes or only x or y specifically.

base_theme
The **ggplot2** theme on which to build the axis-zero graph. Note that if this is not ggplot2::theme_void, you will likely have to do a lot of adjustments in ggplot2::theme() afterwards to make it look good.

base_size
Base font size

base_family
Base font family

base_line_size
Base size for line elements

base_rect_size
size for rect elements

Details

Note that this breaks connection with ggplot things like axis titles and axis range, which will now require considerable fiddling.

Only works if both x and y axes are numeric.

digits = 0,

Examples

ggplot(iris,
aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point() +
  axis_zero(x_title = 'Sepal.Width',
            y_title = 'Sepal.Length',
            y_max = 8, x_max = 5,
            title_distance = .6, digits = 1)

#> Error in ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)): could not find function "ggplot"

# A little messy - data overlaps axis labels. But a four-quadrant graph!

ggplot(iris,
aes(x = Sepal.Width - mean(Sepal.Width), y = Sepal.Length - mean(Sepal.Length))) +
  geom_point() +
  axis_zero(x_title = 'Sepal.Width',
            y_title = 'Sepal.Length',
            y_max = 10, x_max = 3,, x_min = -3,
            quadrants = 1:4,
            title_distance_x = 1.2, title_distance_y = .5, digits = 1,
            axis_distance_x = .4, axis_opts = list(color = 'red'))

#> Error in ggplot(iris, aes(x = Sepal.Width - mean(Sepal.Width), y = Sepal.Length -     mean(Sepal.Length))): could not find function "ggplot"