Print a vector formatted as valid R code for a vector

This function takes a string, such as "a b c", and returns code that will produce that vector in R, such as c("a","b","c").

c_ify(
  s,
  from = c("r", "r_char", "stata", "python", "py_char"),
  from_start = NULL,
  from_sep = NULL,
  from_end = NULL,
  string = from %in% c("r_char", "py_char", "stata"),
  to_start = NULL,
  to_sep = NULL,
  to_end = NULL,
  trim = TRUE,
  silent = FALSE
)

Arguments

s
A string

from
The formatting of the origin string. This can be 'stata' or 'r' (a b c), 'r_char' ("a" "b" "c"), 'python' ([a, b, c]), or 'py_char' (['a', 'b', 'c']). Any format can be supported using custom from_ options below.

from_start
Text at the beginning of the string to remove before separating the elements. Or just give a number to remove that many characters.

from_sep
Character(s) separating each element.

from_end
Text at the end of the string to remove before separating the elements. Or just give a number to remove that many characters.

string
Set to TRUE to set the to_ defaults to produce an R character vector, or FALSE to set the to_ defaults to produce an R numeric vector.

to_start
Text at the beginning of the output string.

to_sep
Character(s) separating each element in the output.

to_end
Text at the end of the output string.

trim
Merge all multiple-whitespaces into single whitespaces before separating the elements, handy if from_sep = ' '.

silent
Suppress the message to console and just return the result as a string.

Details

This is handy when copy/pasting vectors from R console output or other languages (like Stata) into your code.

Examples

c_ify('1 2 3')

#> c("1", "2", "3")

#> [1] "c(\"1\", \"2\", \"3\")"

c_ify('[1, 2, 3]', from = 'python')

#> Error in UseMethod("type"): no applicable method for 'type' applied to an object of class "NULL"

c_ify('var1 var2 var3', from = 'stata')

#> c("var1", "var2", "var3")

#> [1] "c(\"var1\", \"var2\", \"var3\")"

c_ify('%%%a_b_c$$$', from_start = '%%%', from_sep = '_', from_end = '$$$', string = TRUE)

#> c("%%%a", "b", "c$$$")

#> [1] "c(\"%%%a\", \"b\", \"c$$$\")"